+++ Sources/PicshopIntent/RawIntentStep.swift
@@ -200,6 +200,7 @@ public enum IntentNormalizer {
             "cutwords": .cutWords, "removewords": .cutWords, "deletewords": .cutWords, "cutphrase": .cutWords, "cuttext": .cutWords, "textcut": .cutWords,
             "beatsync": .syncToBeat, "cuttobeat": .syncToBeat, "synctomusic": .syncToBeat, "reframe": .smartReframe, "autoreframe": .smartReframe,
             "voiceisolation": .enhanceVoice, "isolatevoice": .enhanceVoice, "cleanaudio": .enhanceVoice, "denoiseaudio": .enhanceVoice,
+            "autocrop": .autoCrop, "bestcrop": .autoCrop, "smartcrop": .autoCrop,
             "cleanup": .cleanUp, "removedistractions": .cleanUp, "removepassersby": .cleanUp, "removetourists": .cleanUp,
             "textbehind": .textBehind, "textbehindsubject": .textBehind, "deptheffect": .textBehind,
             "moveobject": .moveObject, "magicmove": .moveObject, "shiftobject": .moveObject, "reposition": .moveObject,